Rename wifi_provisioner component to wifi
This component handles everything to connect the AirDAC to wifi. If no stored credentials exist, or they don't work, AirDAC launches a soft access point with a captive portal where the wi-fi credentials can be entered. After they are retrieved, the system restarts and re-attempts to connect to the wi-fi, or repeats above.